+++ app/TypeDef/index.js
@@ -20,6 +20,47 @@ module.exports = (XIBLE, EXPRESS_APP) => {
       }
     }
 
+    /**
+     * Verifies whether the given typeDef matches this typeDef.
+     * If not matched directly, the extends property (-tree) of the given typeDef
+     * is verified against this typeDef.
+     * @returns {Boolean}
+     */
+    matches(typeDef) {
+      if (typeDef === this) {
+        return true;
+      }
+
+      if (!typeDef || !typeDef.extends) {
+        return false;
+      }
+
+      // check for extends
+      if (typeof typeDef.extends === 'string') {
+        const extendsTypeDef = typeDefs[typeDef.extends];
+        if (!extendsTypeDef || extendsTypeDef === typeDef) {
+          return false;
+        }
+
+        return this.matches(extendsTypeDef);
+      }
+
+      if (Array.isArray(typeDef.extends)) {
+        for (let i = 0; i < typeDef.extends.length; i += 1) {
+          const extendsTypeDef = typeDefs[typeDef.extends[i]];
+          if (!extendsTypeDef) {
+            continue;
+          }
+
+          if (this.matches(extendsTypeDef)) {
+            return true;
+          }
+        }
+      }
+
+      return false;
+    }
+
     static getAll() {
       return typeDefs;
     }