+++ nodes/input/input.filter-type/index.js
@@ -1,0 +1,10 @@
+'use strict';
+
+module.exports = (NODE) => {
+  const anyIn = NODE.getInputByName('any');
+
+  const filteredOut = NODE.getOutputByName('filtered');
+  filteredOut.on('trigger', async (conn, state) => {
+    return anyIn.getValues(state);
+  });
+};